refactor: use shadcn Card in remaining components
