+++ src/components/admin/CSVDataInspector.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import Papa from 'papaparse';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 import Button from '../ui/Button';
 import { AlertTriangle, Check, Info } from 'lucide-react';
