+++ src/components/admin/DataImportPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import MainLayout from '../layout/MainLayout';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 import Button from '../ui/Button';
 import Input from '../ui/Input';
 import Select from '../ui/Select';
