+++ src/components/admin/SimpleImportTool.tsx
@@ -9,7 +9,7 @@ import {
   Info, 
   Database
 } from 'lucide-react';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 import Button from '../ui/Button';
 import Select from '../ui/Select';
 import { useData } from '../../context/DataContext';
