+++ src/components/contacts/ContactForm.tsx
@@ -18,7 +18,7 @@ import {
   ArrowLeft,
   Save
 } from 'lucide-react';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 import Button from '../ui/Button';
 import Input from '../ui/Input';
 import Select from '../ui/Select';
