+++ src/components/document-templates/DocumentGenerator.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 
 /**
  * Placeholder DocumentGenerator component
