+++ src/components/invoices/InvoiceDetail.tsx
@@ -3,7 +3,7 @@ import { useParams, useNavigate } from 'react-router-dom';
 import { useData } from '../../context/DataContext';
 import { format, parseISO, isValid } from 'date-fns';
 import { ArrowLeft, Plus } from 'lucide-react';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 import Button from '../ui/Button';
 import Table from '../ui/Table';
 import InvoiceForm from './InvoiceForm';
