+++ src/components/notifications/NotificationSettingsForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 
 /**
  * Placeholder NotificationSettingsForm component
