+++ src/components/templates/PageTemplates.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Typography from '../ui/Typography';
-import Card from '../ui/Card';
+import { Card } from '../ui/shadcn-card';
 import Button from '../ui/Button';
 import Input from '../ui/Input';
 import Select from '../ui/Select';
